// src/types/config.ts
export interface ConfigData {
  id?: number;
  key: string;
  value: string;
  description?: string;
}

export interface CodeDetailData {
  id?: number;
  codeId: number;
  name: string;
  description: string;
  value: string;
  lastUpdatedAt?: string;
}